Classify into three package: Main, Task, Command
