+++ src/main/java/Command/Command.java
@@ -1,4 +1,9 @@
-import java.io.IOException;
+package Command;
+
+import Main.TaskList;
+import Main.Storage;
+import Main.DukeException;
+import Main.Ui;
 
 public abstract class Command {
