+++ src/main/java/Main/DukeException.java
@@ -1,5 +1,7 @@
+package Main;
+
 public class DukeException extends Exception{
-    String message;
+    private String message;
 
     public DukeException(String message) {
         this.message = message;
