+++ src/main/java/Main/Storage.java
@@ -1,9 +1,16 @@
+package Main;
+
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Scanner;
+
+import Task.Task;
+import Task.Event;
+import Task.Deadline;
+import Task.Todo;
 
 public class Storage {
     private String filePath;
