+++ src/main/java/Main/Ui.java
@@ -1,4 +1,8 @@
+package Main;
+
 import java.util.Scanner;
+
+import Task.Task;
 
 public class Ui {
 
@@ -12,7 +16,7 @@ public class Ui {
 
     public static void greet() {
         System.out.println(newLine);
-        System.out.println(indentation + "Hello! I'm Duke");
+        System.out.println(indentation + "Hello! I'm Main.Duke");
         System.out.println(indentation + "What can I do for you?");
         System.out.println(newLine);
     }
