+++ src/main/java/Task/Todo.java
@@ -1,6 +1,9 @@
-import java.time.LocalDate;
+package Task;
 
-public class Todo extends Task{
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+
+public class Todo extends Task {
 
     public Todo(String name) {
         super(name);
